FINERACT-959 make javac fail for Xlint:cast

+++ fineract-provider/src/integrationTest/java/org/apache/fineract/integrationtests/common/CenterDomain.java
@@ -242,6 +242,6 @@ public class CenterDomain implements Comparable<CenterDomain> {
 
     @Override
     public int compareTo(CenterDomain cd) {
-        return ((Integer) this.id).compareTo(cd.getId());
+        return ( this.id).compareTo(cd.getId());
     }
 }
